Refactor mocha
* Refactored the utilities into its own header files.
* End users no long need to use a lambda function to compare two objects of the same type.
